TESTS: consolidate polars get_x_y tests into test_datasets
The get_x_y drivers assert container types differently per backend, so
the polars cases move as a dedicated factory next to the pandas one,
reusing the shared data helpers. Collected test count (68) and coverage
are unchanged.
